Add a build system and nested project for testing
Previously we would only load libraries shipped with the operating system
during testing, now we can build our own.

This commit adds an internal library named examplelib and a gradle build
system which allows the library to be built before running tests. This is
useful because it allows us to ship the same library to all operating
systems and test how it is loaded, rather than only using system libraries
in tests.