Issue #10 -- unix lib.rs passes non zero terminated strings to dlopen
Converting to CString type before passing as a C pointer to dlopen added the \0 termination and obviates one of the casts. Tests would fail if the same rigor was applied to dlsym because optional \0 termination is used as a feature (see example calls of find_func() in lib_impl/lib.rs).